+++ neoforge/src/main/java/com/telepathicgrunt/the_bumblezone/configs/neoforge/BzConfigHandler.java
@@ -1,5 +1,7 @@
 package com.telepathicgrunt.the_bumblezone.configs.neoforge;
 
+import com.electronwill.nightconfig.core.file.CommentedFileConfig;
+import com.electronwill.nightconfig.core.io.WritingMode;
 import com.telepathicgrunt.the_bumblezone.items.potions.ConfigSafePotion;
 import com.telepathicgrunt.the_bumblezone.modinit.BzPotions;
 import net.neoforged.bus.api.IEventBus;
@@ -7,26 +9,26 @@ import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.config.IConfigSpec;
 import net.neoforged.fml.config.ModConfig;
 import net.neoforged.fml.event.config.ModConfigEvent;
-import net.neoforged.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.neoforged.fml.loading.FMLPaths;
+import net.neoforged.neoforge.common.ModConfigSpec;
 
 import java.nio.file.Files;
+import java.nio.file.Path;
 
 public class BzConfigHandler {
 
-    public static void setup() {
-        IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
-        bus.addListener(BzConfigHandler::onConfigLoad);
-        bus.addListener(BzConfigHandler::onConfigReload);
+    public static void setup(IEventBus modEventBus) {
+        modEventBus.addListener(BzConfigHandler::onConfigLoad);
+        modEventBus.addListener(BzConfigHandler::onConfigReload);
 
         try {
             Files.createDirectories(FMLPaths.CONFIGDIR.get().resolve("the_bumblezone"));
-            ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, BzClientConfig.GENERAL_SPEC, "the_bumblezone/client.toml");
-            ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, BzGeneralConfig.GENERAL_SPEC, "the_bumblezone/general.toml");
-            ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, BzWorldgenConfig.GENERAL_SPEC, "the_bumblezone/worldgen.toml");
-            ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, BzDimensionConfig.GENERAL_SPEC, "the_bumblezone/dimension.toml");
-            ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, BzBeeAggressionConfig.GENERAL_SPEC, "the_bumblezone/bee_aggression.toml");
-            ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, BzModCompatibilityConfig.GENERAL_SPEC, "the_bumblezone/mod_compatibility.toml");
+            createAndLoadConfigs(ModConfig.Type.CLIENT, BzClientConfig.GENERAL_SPEC, "the_bumblezone/client.toml");
+            createAndLoadConfigs(ModConfig.Type.COMMON, BzGeneralConfig.GENERAL_SPEC, "the_bumblezone/general.toml");
+            createAndLoadConfigs(ModConfig.Type.COMMON, BzWorldgenConfig.GENERAL_SPEC, "the_bumblezone/worldgen.toml");
+            createAndLoadConfigs(ModConfig.Type.COMMON, BzDimensionConfig.GENERAL_SPEC, "the_bumblezone/dimension.toml");
+            createAndLoadConfigs(ModConfig.Type.COMMON, BzBeeAggressionConfig.GENERAL_SPEC, "the_bumblezone/bee_aggression.toml");
+            createAndLoadConfigs(ModConfig.Type.COMMON, BzModCompatibilityConfig.GENERAL_SPEC, "the_bumblezone/mod_compatibility.toml");
         }
         catch (Exception e) {
             throw new RuntimeException("Failed to create Bumblezone config files: ", e);
@@ -55,4 +57,17 @@ public class BzConfigHandler {
         if (spec == BzModCompatibilityConfig.GENERAL_SPEC) BzModCompatibilityConfig.copyToCommon();
     }
 
+    private static void createAndLoadConfigs(ModConfig.Type type, ModConfigSpec spec, String path) {
+        ModLoadingContext.get().registerConfig(type, spec, path);
+
+        final CommentedFileConfig configData = CommentedFileConfig.builder(FMLPaths.CONFIGDIR.get().resolve(path))
+                                                .preserveInsertionOrder()
+                                                .autoreload()
+                                                .writingMode(WritingMode.REPLACE)
+                                                .sync()
+                                                .build();
+
+        configData.load();
+        spec.setConfig(configData);
+    }
 }